+++ typescript/aleo-sdk/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @hyperlane-xyz/aleo-sdk
 
+## 38.0.0
+
+### Patch Changes
+
+- Updated dependencies [961a89d]
+  - @hyperlane-xyz/provider-sdk@7.2.0
+  - @hyperlane-xyz/utils@38.0.0
+
+## 37.0.0
+
+### Minor Changes
+
+- 9c8b435: Added altvm support in warp check
+
+### Patch Changes
+
+- 3771b2b: Fixed aleo warp send with retries
+- Updated dependencies [df34a68]
+- Updated dependencies [cc4bdb6]
+- Updated dependencies [31f8b51]
+- Updated dependencies [97e8ca1]
+  - @hyperlane-xyz/provider-sdk@7.1.0
+  - @hyperlane-xyz/utils@37.0.0
+
 ## 36.0.0
 
 ### Major Changes